feat: add LAGON_WORKERS env variable
