+++ crates/serverless/src/worker.rs
@@ -5,10 +5,11 @@ use lagon_runtime_isolate::{options::IsolateOptions, Isolate};
 use lagon_runtime_utils::Deployment;
 use log::{error, info};
 use metrics::{decrement_gauge, histogram, increment_gauge};
+use rand::{Rng, SeedableRng};
 use tokio::sync::RwLock;
 use tokio_util::task::LocalPoolHandle;
 
-use crate::{POOL_SIZE, REGION, SNAPSHOT_BLOB};
+use crate::{REGION, SNAPSHOT_BLOB, WORKERS};
 
 pub type Workers =
     Arc<RwLock<HashMap<usize, (flume::Sender<WorkerEvent>, flume::Receiver<WorkerEvent>)>>>;
@@ -27,8 +28,41 @@ pub enum WorkerEvent {
     },
 }
 
+pub async fn create_workers() -> Workers {
+    let workers = Arc::new(RwLock::new(HashMap::new()));
+
+    for i in 0..*WORKERS {
+        let worker = flume::unbounded();
+        workers.write().await.insert(i, worker);
+    }
+
+    workers
+}
+
+pub async fn get_thread_id(
+    thread_ids: Arc<RwLock<HashMap<String, usize>>>,
+    hostname: &String,
+) -> usize {
+    let thread_ids_reader = thread_ids.read().await;
+
+    let thread_id = match thread_ids_reader.get(hostname) {
+        Some(thread_id) => *thread_id,
+        None => {
+            let mut rng = rand::rngs::StdRng::from_entropy();
+            let id = rng.gen_range(0..*WORKERS);
+
+            drop(thread_ids_reader);
+
+            thread_ids.write().await.insert(hostname.clone(), id);
+            id
+        }
+    };
+
+    thread_id
+}
+
 pub async fn start_workers(workers: Workers) {
-    let pool = LocalPoolHandle::new(POOL_SIZE);
+    let pool = LocalPoolHandle::new(*WORKERS);
 
     for (id, worker) in workers.read().await.iter() {
         let id = *id;
